+++ pages/department/departments-list.php
@@ -65,7 +65,7 @@ $data = select($query);
                                                 </a>
                                             </td>
                                             <td class="text-left">
-                                                <a href="../../api/departments/deleteDepartment.php?id=<?=$row['Id']?>" class="btn btn-sm bg-danger-light ml-2">
+                                                <a href="../../api/departments/deleteDepartment.php?id=<?=$row['Id']?>" onclick="return confirm('Are you sure to delete this department')" class="btn btn-sm bg-danger-light ml-2">
                                                     <i class="fas fa-trash"></i>
                                                 </a>
                                             </td>
